SRCH-6530 added metadata default to transform functions.  Reorganized functions in transform to be related to creating the documents and moved other things to helpers or parse
