+++ search_gov_crawler/indexing/parse.py
@@ -1,6 +1,15 @@
+import logging
+import re
+from datetime import datetime, timedelta, timezone
+from typing import Any
+
+from pypdf import PageObject, PdfReader
+from pypdf.generic import IndirectObject
 from scrapy import Selector
 
 from search_gov_crawler.search_gov_spiders.helpers import content
+
+log = logging.getLogger(__name__)
 
 
 def extract_article_content(html_selector: Selector) -> str:
@@ -25,7 +34,7 @@ def extract_article_content(html_selector: Selector) -> str:
     return content.replace_whitespace(content_text)
 
 
-def get_meta_values(html_selector: Selector, meta_names: list) -> dict:
+def get_html_meta(html_selector: Selector, meta_names: list) -> dict:
     """
     Extracts meta tag values by their name or property attributes and returns a dictionary.
 
@@ -50,7 +59,7 @@ def convert_html_scrapy(html_content: str) -> dict:
     return_obj = {}
     html_selector = Selector(text=html_content)
 
-    meta_tags = get_meta_values(
+    meta_tags = get_html_meta(
         html_selector,
         [
             "keywords",
@@ -99,3 +108,145 @@ def convert_html_scrapy(html_content: str) -> dict:
     return_obj["content"] = extract_article_content(html_selector)
 
     return return_obj
+
+
+def get_pdf_text(reader: PdfReader) -> tuple[str, list[tuple[str, PageObject]]]:
+    """
+    Returns clean text/content from all pdf pages
+
+    Args:
+        reader: PdfReader from pypdf
+
+    Returns:
+        (string) without new any special characters
+    """
+    text = ""
+    pages = []
+    for page in reader.pages:
+        page_text = page.extract_text()
+        text += page_text + " "
+        pages.append((page_text, page))
+    return (text, pages)
+
+
+def get_pdf_meta(reader: PdfReader) -> dict:
+    """
+    Returns pdf metadata as a dict after its been cleaned.
+
+    Args:
+        reader: PdfReader from pypdf
+
+    Returns:
+        metadata object with possible keys: https://exiftool.org/TagNames/PDF.html
+    """
+    if not reader.metadata:
+        return {}
+
+    clean_metadata = {}
+    for k, v in reader.metadata.items():
+        resolved_value = v.get_object() if isinstance(v, IndirectObject) else v
+        clean_metadata[str(k).removeprefix("/")] = parse_exif_date(resolved_value, apply_tz_offset=False)
+
+    return clean_metadata
+
+
+def parse_exif_date(value: Any, *, apply_tz_offset: bool) -> Any:
+    """
+    Parses a value as date if matched the conventional pdf/exif date format. If parsing fails,
+    returns the original value
+
+    Examples of str date format:
+        "D:20150113143419Z00'00'"
+        "D:20191018122555-04'00'"
+        "D:20191018162538"
+
+    Args:
+        value: The value to parse.
+
+    Returns:
+        A datetime.datetime object if parsing is successful, otherwise the original value.
+    """
+    if not isinstance(value, str):
+        return value
+
+    if value.startswith("D:"):
+        date_string = value.removeprefix("D:")
+
+        proper_date_format = re.match(
+            r"^(\d{4})(\d{2})(\d{2})(\d{2})?(\d{2})?(\d{2})?([+\-Z]{0,1})?(\d{2})?'?(\d{2})?'?$",
+            date_string,
+        )
+        misformed_date_format = re.match(r"^[0-9zZ+\-']*$", date_string)
+
+        if proper_date_format:
+            year = int(proper_date_format.group(1))
+            month = int(proper_date_format.group(2))
+            day = int(proper_date_format.group(3))
+            hour = int(proper_date_format.group(4)) if proper_date_format.group(4) else 0
+            minute = int(proper_date_format.group(5)) if proper_date_format.group(5) else 0
+            second = int(proper_date_format.group(6)) if proper_date_format.group(6) else 0
+            tz_sign = proper_date_format.group(7) or "Z"
+            tz_hour = int(proper_date_format.group(8)) if proper_date_format.group(8) else 0
+            tz_minute = int(proper_date_format.group(9)) if proper_date_format.group(9) else 0
+
+            # Handle timezone offset if matched
+            if proper_date_format.group(7) and apply_tz_offset:
+                tz_multiplier = -1 if tz_sign == "-" else 1
+                offset = timedelta(hours=tz_hour, minutes=tz_minute) * tz_multiplier
+                tz = timezone(offset=offset)
+            else:
+                tz = None
+
+            try:
+                return datetime(year, month, day, hour, minute, second, tzinfo=tz)
+            except ValueError:
+                log.debug("Failed to parse date string: %s", value)
+                return None
+        elif misformed_date_format:
+            log.debug("Failed to parse date string: %s", value)
+            return None
+        else:
+            pass  # Starts with D: but probably not a date
+
+    return content.sanitize_text(value)
+
+
+def get_pdf_links(pages: list[tuple[str, PageObject]]):
+    """
+    Returns a set of links for all pages in the PDF
+
+    Args:
+        pages: list of tuples containing (text, PageObject)
+
+    Returns:
+        (list[str]) unique set of links
+    """
+    key = "/Annots"
+    uri = "/URI"
+    ank = "/A"
+    links = set()  # Use a set for unique links
+
+    for page_item in pages:
+        text, page = page_item
+        # Get all visible links from text
+        page_links = re.findall(r"https?://\S+|www\.\S+", text)
+        for link in page_links:
+            links.add(link)
+
+        # Get all hidden links from annotations
+        page_object = page.get_object()
+        if key in page_object.keys():  # noqa: SIM118
+            ann = page_object[key]
+            for a in ann:
+                u = a.get_object()
+                try:
+                    if ank in u and uri in u[ank].keys():  # noqa: SIM118
+                        link = u[ank][uri]
+                        # Convert bytes to string if necessary
+                        if isinstance(link, bytes):
+                            link = link.decode("utf-8")
+                        links.add(link)
+                except ValueError:
+                    pass
+
+    return list(links)
